Add go playground as run possibility
Beside installing Go on the own machine, it is possible to execute all the
example programs within the playground.
The advantage is that it is more convient to execute the examples for users who
are either have not installed Go or are not willing to clone the repository to
execute the examples.